Re-implement ep_in_complete using TXINI flag
Based on the usage in usb_device's control
pipe, I think ep_in_complete is supposed to
signal when the tx buffer is free so that more
can be written, and that is what TXINI signals.

Still keeping the pending_ins bitmask, as it
still should only be reported once.